+++ CHANGES.rst
@@ -15,6 +15,7 @@ Development
 - Try 11 dB attenuation for measuring vcc
 - Disable heartbeat through RGB-LED, just blink twice on startup
 - Turn off interrupts when powering down the HX711
+- Improve inline documentation and logging
 
 
 2019-06-17 0.4.0